gpu/capture: give the file format its own folder
The two entry points were still one 2,843-line file. They are one contract read from
opposite ends -- a field written and not read is silent data loss, read and not written is
a parse that succeeds on garbage -- so they belong beside each other, not inside each
other.

  capture/serialize/capture_codecs.hpp       422   cursors + per-field codecs
  capture/serialize/capture_serialize.cpp    1055   serialize_gpu_capture
  capture/serialize/capture_deserialize.cpp  1479   deserialize_gpu_capture + legacy restoration

A pure move: git mv plus two rewritten includes, by move_module.py, so the diff is
relocation rather than retyped code. The split before it was proven by reconstruction --
the two outputs rebuild the original 132,740 bytes exactly.

The folder carries an AGENTS.md saying what the boundary against its parent IS: capture/
decides what to record, serialize/ decides only how it is written down, and the line
between them is the version contract. It also records the two rules that are invisible in
the code -- a version bump is append-only because the cursors are positional, and every
count read from a header is bounds-checked before it sizes an allocation, because these
are the files that parse bytes off disk.

Builds clean; ctest --no-tests=error -j4 reports 457/457, exit 0.
